+++ ams-webapi-server/src/main/java/com/precisionhawk/ams/webservices/impl/WorkOrderWebServiceImpl.java
@@ -1,16 +1,19 @@
 package com.precisionhawk.ams.webservices.impl;
 
 import com.precisionhawk.ams.bean.WorkOrderSearchParams;
+import com.precisionhawk.ams.bean.security.ServicesSessionBean;
 import com.precisionhawk.ams.dao.DaoException;
 import com.precisionhawk.ams.dao.WorkOrderDao;
 import com.precisionhawk.ams.domain.WorkOrder;
 import com.precisionhawk.ams.webservices.WorkOrderWebService;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.ws.rs.BadRequestException;
 import javax.ws.rs.InternalServerErrorException;
+import javax.ws.rs.NotAuthorizedException;
 
 /**
  *
@@ -23,9 +26,10 @@ public class WorkOrderWebServiceImpl extends AbstractWebService implements WorkO
 
     @Override
     public WorkOrder retrieveById(String authToken, String orderNumber) {
+        ServicesSessionBean sess = super.lookupSessionBean(authToken);
         ensureExists(orderNumber, "Work order number is required.");
         try {
-            return dao.retrieveById(orderNumber);
+            return authorize(sess, validateFound(dao.retrieveById(orderNumber)));
         } catch (DaoException ex) {
             throw new InternalServerErrorException(String.format("Error retrieving work order for order number %s.", orderNumber), ex);
         }
@@ -34,9 +38,11 @@ public class WorkOrderWebServiceImpl extends AbstractWebService implements WorkO
     @Override
     public List<WorkOrder> search(String authToken, WorkOrderSearchParams searchBean) {
         ensureExists(searchBean, "Search parameters are required.");
+        ServicesSessionBean sess = super.lookupSessionBean(authToken);
+        authorize(sess, searchBean);
         if (searchBean.hasCriteria()) {
             try {
-                return dao.search(searchBean);
+                return cleanseUnauthorizedWorkOrders(sess, dao.search(searchBean));
             } catch (DaoException ex) {
                 throw new InternalServerErrorException("Error searching for work orders.");
             }
@@ -47,7 +53,9 @@ public class WorkOrderWebServiceImpl extends AbstractWebService implements WorkO
 
     @Override
     public WorkOrder create(String authToken, WorkOrder workOrder) {
+        ServicesSessionBean sess = super.lookupSessionBean(authToken);
         ensureExists(workOrder, "Work order is required.");
+        authorize(sess, workOrder);
         if (workOrder.getOrderNumber() == null) {
             workOrder.setOrderNumber(UUID.randomUUID().toString().split("-")[0]);
         }
@@ -64,9 +72,14 @@ public class WorkOrderWebServiceImpl extends AbstractWebService implements WorkO
 
     @Override
     public void delete(String authToken, String orderNumber) {
+        ServicesSessionBean sess = super.lookupSessionBean(authToken);
         ensureExists(orderNumber, "Work order number is required.");
         try {
-            dao.delete(orderNumber);
+            WorkOrder order = dao.retrieveById(orderNumber);
+            if (order != null) {
+                authorize(sess, order);
+                dao.delete(orderNumber);
+            }
         } catch (DaoException ex) {
             throw new InternalServerErrorException(String.format("Error deleting work order for order number %s.", orderNumber), ex);
         }
@@ -74,17 +87,52 @@ public class WorkOrderWebServiceImpl extends AbstractWebService implements WorkO
 
     @Override
     public void update(String authToken, WorkOrder workOrder) {
+        ServicesSessionBean sess = super.lookupSessionBean(authToken);
         ensureExists(workOrder, "Work order is required.");
         ensureExists(workOrder.getOrderNumber(), "Work order number is required.");
-        if (workOrder.getOrderNumber() == null) {
-            workOrder.setOrderNumber(UUID.randomUUID().toString().split("-")[0]);
-        }
+        authorize(sess, workOrder);
         try {
-            if (!dao.update(workOrder)) {
-                throw new BadRequestException(String.format("The work order %s already exists.", workOrder.getOrderNumber()));
+            boolean updated = false;
+            WorkOrder order = dao.retrieveById(workOrder.getOrderNumber());
+            if (order != null) {
+                authorize(sess, order);
+                updated = dao.update(workOrder);
+            }
+            if (!updated) {
+                throw new BadRequestException(String.format("The work order %s does not already exist.", workOrder.getOrderNumber()));
             }
         } catch (DaoException ex) {
             throw new InternalServerErrorException("Error storing new work order", ex);
         }
-    }    
+    }
+    
+    private boolean testAuthorization(ServicesSessionBean sess, WorkOrder order, String ... groupKeys) {
+        for (String siteId : order.getSiteIds()) {
+            if (sess.getCredentials().checkAuthorization(null, null, siteId, false, groupKeys)) {
+                return true;
+            }
+        }
+        return false;
+    }
+    
+    private WorkOrder authorize(ServicesSessionBean sess, WorkOrder order, String ... groupKeys) {
+        if (!sess.isTokenValid()) {
+            throw new NotAuthorizedException(sess.getReason());
+        }
+        if (testAuthorization(sess, order, groupKeys)) {
+            return order;
+        } else {
+            throw new NotAuthorizedException(String.format("User not authorized for work order %s", order.getOrderNumber()));
+        }
+    }
+    
+    private List<WorkOrder> cleanseUnauthorizedWorkOrders(ServicesSessionBean sess, List<WorkOrder> input, String ... groupKeys) {
+        List<WorkOrder> output = new ArrayList<>(input.size());
+        for (WorkOrder order : input) {
+            if (testAuthorization(sess, order, groupKeys)) {
+                output.add(order);
+            }
+        }
+        return output;
+    }
 }
